pyproject.toml
[project]
name = "llm-observability"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai>=2.28.0",
    "openinference-instrumentation-openai>=0.1.42",
]
